feat: refactor bundle component naming
